Move debug transforms to module
